[IMP] Started splitting backend tests according to topics
